+++ src/main/java/org/thingsboard/rule/engine/vix/TbVixClearAlarmsNode.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.rule.engine.node.action;
+package org.thingsboard.rule.engine.vix;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.util.concurrent.Futures;
