+++ src/main/java/org/thingsboard/rule/engine/vix/TbVixFindEntityNode.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.rule.engine.node.transform;
+package org.thingsboard.rule.engine.vix;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import lombok.extern.slf4j.Slf4j;
