+++ src/main/java/org/thingsboard/rule/engine/vix/VixEntity.java
@@ -13,22 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.rule.engine.node.action;
+package org.thingsboard.rule.engine.vix;
 
-import lombok.Data;
-import org.thingsboard.rule.engine.api.NodeConfiguration;
-
-@Data
-public class TbVixAlarmsNodeConfiguration implements NodeConfiguration<TbVixAlarmsNodeConfiguration> {
-
-    private int evaluatedDataPoints;
-    private int dataPointsToAlarm;
-
-    @Override
-    public TbVixAlarmsNodeConfiguration defaultConfiguration() {
-        TbVixAlarmsNodeConfiguration configuration = new TbVixAlarmsNodeConfiguration();
-        configuration.setEvaluatedDataPoints(5);
-        configuration.setDataPointsToAlarm(5);
-        return configuration;
-    }
+public enum VixEntity {
+    IP, TM, OTHER, CUSTOMER
 }
